+++ GranularSoundControl/GranularSoundControlInfo.cs
@@ -1,0 +1,7 @@
+﻿namespace GranularSoundControl; 
+
+public static class GranularSoundControlInfo {
+    public const string Guid = "GranularSoundControl";
+    public const string Name = "GranularSoundControl";
+    public const string Version = "1.4.0";
+}